fixup gdb for unrolled main loop
