+++ components/blackmagic/exception.c
