+++ components/blackmagic/gdb_if.h
@@ -21,11 +21,6 @@
 #ifndef __GDB_IF_H
 #define __GDB_IF_H
 
-/*#if PC_HOSTED == 0
-#include <libopencm3/usb/usbd.h>
-void gdb_usb_out_cb(usbd_device *dev, uint8_t ep);
-#endif*/
-
 int gdb_if_init(void);
 unsigned char gdb_if_getchar(void);
 unsigned char gdb_if_getchar_to(int timeout);
