+++ sdkconfig.defaults
@@ -1,4 +1,46 @@
 # This file was generated using idf.py save-defconfig. It can be edited manually.
 # Espressif IoT Development Framework (ESP-IDF) Project Minimal Configuration
 #
-CONFIG_IDF_TARGET="esp32s3"
+CONFIG_BOOTLOADER_APP_ROLLBACK_ENABLE=y
+CONFIG_ESPTOOLPY_FLASHSIZE_8MB=y
+CONFIG_ESPTOOLPY_HEADER_FLASHSIZE_UPDATE=y
+CONFIG_PARTITION_TABLE_CUSTOM=y
+CONFIG_COMPILER_OPTIMIZATION_PERF=y
+CONFIG_COMPILER_STACK_CHECK_MODE_STRONG=y
+CONFIG_APPTRACE_DEST_JTAG=y
+CONFIG_APPTRACE_DEST_USB_CDC=y
+CONFIG_SPI_MASTER_IN_IRAM=y
+CONFIG_UART_ISR_IN_IRAM=y
+CONFIG_GPIO_CTRL_FUNC_IN_IRAM=y
+CONFIG_GPTIMER_CTRL_FUNC_IN_IRAM=y
+CONFIG_GPTIMER_ISR_IRAM_SAFE=y
+CONFIG_HTTPD_MAX_REQ_HDR_LEN=1024
+CONFIG_HTTPD_WS_SUPPORT=y
+CONFIG_ESP_SLEEP_POWER_DOWN_FLASH=y
+CONFIG_PM_ENABLE=y
+CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ_240=y
+CONFIG_ESP32S3_INSTRUCTION_CACHE_32KB=y
+CONFIG_ESP32S3_DATA_CACHE_64KB=y
+CONFIG_ESP32S3_DATA_CACHE_LINE_64B=y
+CONFIG_ESP_MAIN_TASK_AFFINITY_NO_AFFINITY=y
+CONFIG_ESP_TASK_WDT_PANIC=y
+CONFIG_ESP_PANIC_HANDLER_IRAM=y
+CONFIG_ESP_DEBUG_STUBS_ENABLE=y
+CONFIG_ESP_WIFI_SLP_IRAM_OPT=y
+CONFIG_ESP_WIFI_STA_DISCONNECTED_PM_ENABLE=y
+CONFIG_FREERTOS_THREAD_LOCAL_STORAGE_POINTERS=4
+CONFIG_FREERTOS_VTASKLIST_INCLUDE_COREID=y
+CONFIG_FREERTOS_GENERATE_RUN_TIME_STATS=y
+CONFIG_FREERTOS_WATCHPOINT_END_OF_STACK=y
+CONFIG_HEAP_POISONING_LIGHT=y
+CONFIG_LWIP_LOCAL_HOSTNAME="farpatch"
+CONFIG_LWIP_IRAM_OPTIMIZATION=y
+CONFIG_LWIP_MAX_SOCKETS=16
+CONFIG_LWIP_DHCP_RESTORE_LAST_IP=y
+CONFIG_LWIP_AUTOIP=y
+# CONFIG_MQTT_PROTOCOL_311 is not set
+# CONFIG_MQTT_TRANSPORT_SSL is not set
+# CONFIG_MQTT_TRANSPORT_WEBSOCKET is not set
+CONFIG_NEWLIB_NANO_FORMAT=y
+# CONFIG_WPA_MBEDTLS_TLS_CLIENT is not set
+CONFIG_DEFAULT_AP_PASSWORD="12345678"
